# server/docs/swagger.yaml
definitions:
  request.UserLoginRequest:
    properties:
      loginId:
        type: string
      password:
        type: string
    required:
    - loginId
    - password
    type: object
  request.UserRegisterRequest:
    properties:
      loginId:
        type: string
      mail:
        type: string
      name:
        type: string
      password:
        type: string
      rePassword:
        type: string
    required:
    - loginId
    - mail
    - name
    - password
    - rePassword
    type: object
  response.LoginResponse:
    properties:
      jwt:
        type: string
    type: object
info:
  contact: {}
paths:
  /article/list:
    get:
      consumes:
      - application/json
      description: 执行登录返回token
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Response
          schema:
            type: string
      summary: 获取文章列表
      tags:
      - 文章
  /user/login:
    post:
      consumes:
      - application/json
      description: 执行登录返回token
      parameters:
      - description: 登录请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/request.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Response
          schema:
            $ref: '#/definitions/response.LoginResponse'
      summary: 登录
      tags:
      - 用户
  /user/register:
    post:
      consumes:
      - application/json
      description: 用户注册并执行自动登录
      parameters:
      - description: 注册请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/request.UserRegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Response
          schema:
            $ref: '#/definitions/response.LoginResponse'
      summary: 注册
      tags:
      - 用户
swagger: "2.0"